requirements.txt
flask
mysql-connector-python
openpyxl
gunicorn